libsks: get cryptoki return value from the TA
TA can generate a SKS return value together with the GPD TEE return
code. It such case, the GPD TEE return code is TEEC_SUCCESS and the
SKS return code can be read from the 1st invocation parameter which
is defined as an input/output memory reference.

The SKS return code can provide finer information on the reason of
failure when processing the client request.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>